Devlens/backend/requirements.txt
annotated-types==0.7.0
google-generativeai==0.3.2
anyio==4.11.0
beautifulsoup4==4.14.2
certifi==2025.10.5
click==8.3.0
colorama==0.4.6
distro==1.9.0
docstring_parser==0.17.0
exceptiongroup==1.3.0
fastapi==0.118.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.11.0
lxml==6.0.2
pydantic==2.11.10
pydantic-settings==2.11.0
pydantic_core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.20
sniffio==1.3.1
soupsieve==2.8
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
